fix(gv-code): define extension of files to load for a specific CodeMirror mode
An import without a file extension is making webpack (or any other bundler) trying to require all the files matching the name.
Everything is ok with mode containing only a JS file in their folder.
But a few of them also have CSS or HTML files causing issues when the bundler is trying to load them.

Using the JS extension ensure we are only loading the file we have in mind.