feat: add new gv-shema-form-group
This component come from gv-shema-form but without submit / cancel feature.
The goal of this one is to be used by angular inside reactive form (simply)

Component with medium term goal. The idea is to use it for all schema form needs in Gravitee console during the angularJs migration. Until it is replaced by native angular one