fix: dispatch input event to work with Angular's ngDefaultControl
In some cases like copy and paste or select angular does not detect any change because it is waiting for an "input" event.
Yes it's a fix on a hack already in place and commonly used in the portal.
For the console this is temporary and soon replaced by native angular components